requirements.txt
jsonschema>=4.22.0
rfc3339-validator
rfc3987
odfpy